% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_priorid.R
\name{get_priorid}
\alias{get_priorid}
\title{Informative bayesian prior lookup for heterogenity in meta-analyses}
\usage{
get_priorid()
}
\description{
Informative bayesian prior lookup for heterogenity in meta-analyses
}
